+++ src/main/java/fr/tropweb/miningmanager/engine/Engine.java
@@ -7,8 +7,8 @@ import fr.tropweb.miningmanager.commands.Reload;
 import fr.tropweb.miningmanager.commands.Scan;
 import fr.tropweb.miningmanager.commands.struct.CommandManager;
 import fr.tropweb.miningmanager.commands.struct.SubCommand;
+import fr.tropweb.miningmanager.dao.sqlite.SQLiteDAO;
 import fr.tropweb.miningmanager.data.Settings;
-import fr.tropweb.miningmanager.data.sqlite.SQLiteDAO;
 import fr.tropweb.miningmanager.pojo.PlayerLite;
 import lombok.Data;
 import org.bukkit.Server;
@@ -34,6 +34,8 @@ public final class Engine {
     private final ChunkEngine chunkEngine;
     private final BlockEngine blockEngine;
     private final PlayerEngine playerEngine;
+
+    private final MiningEngine miningEngine;
     private final RegenerationEngine regenerationEngine;
 
 
@@ -48,6 +50,7 @@ public final class Engine {
         this.blockEngine = new BlockEngine(this);
         this.playerEngine = new PlayerEngine(this);
 
+        this.miningEngine = new MiningEngine(this);
         this.regenerationEngine = new RegenerationEngine(this);
 
         commands.clear();
@@ -100,10 +103,10 @@ public final class Engine {
         for (final PlayerLite playerLite : this.getPlayerEngine().getPlayerLiteMap().values()) {
 
             // if player has mining task
-            if (playerLite.hasMiningTask()) {
+            if (playerLite.getMiningTask().hasMiningTask()) {
 
                 // stop the mining of all player
-                playerLite.stopMiningTask();
+                playerLite.getMiningTask().stopMiningTask();
 
                 // check if the player exist
                 final Player player = this.getServer().getPlayer(playerLite.getUniqueId());