+++ src/main/java/fr/tropweb/miningmanager/listeners/PlayerEventHandler.java
@@ -1,9 +1,9 @@
 package fr.tropweb.miningmanager.listeners;
 
 import fr.tropweb.miningmanager.Utils;
-import fr.tropweb.miningmanager.commands.threads.MiningThread;
-import fr.tropweb.miningmanager.engine.Engine;
 import fr.tropweb.miningmanager.data.Settings;
+import fr.tropweb.miningmanager.engine.Engine;
+import fr.tropweb.miningmanager.pojo.MiningTask;
 import fr.tropweb.miningmanager.pojo.PlayerLite;
 import org.bukkit.Bukkit;
 import org.bukkit.block.Block;
@@ -45,20 +45,23 @@ public class PlayerEventHandler implements Listener {
         final Block block = event.getClickedBlock();
         if (block == null) return;
 
-        // get player
+        // get player from memory
         final PlayerLite playerLite = this.engine.getPlayerEngine().getPlayerLite(player);
 
+        // get mining task from player
+        final MiningTask miningTask = playerLite.getMiningTask();
+
         // if have the end task and if he doesn't choose chest
-        if (playerLite.hasMiningTask() && !playerLite.hasChooseChest()) {
+        if (miningTask.hasMiningTask() && !miningTask.hasMiningChest()) {
 
             // if it's chest
             if (this.engine.getBlockEngine().isChest(block)) {
 
                 // add chest to player
-                playerLite.setMiningChest(block);
+                miningTask.setMiningChest(block);
 
                 // cancel task
-                playerLite.getBukkitTask().cancel();
+                miningTask.getMiningTask().cancel();
 
                 // get settings
                 final Settings settings = this.engine.getSettings();
@@ -66,12 +69,12 @@ public class PlayerEventHandler implements Listener {
                 // create task
                 final BukkitTask task = Bukkit.getScheduler().runTaskTimer(
                         this.engine.getPlugin(),
-                        () -> new MiningThread(this.engine, player).run(),
+                        () -> this.engine.getMiningEngine().startMining(player),
                         settings.getTickMiningStart(),
                         settings.getTickMiningInterval());
 
                 // save task
-                playerLite.setBukkitTask(task);
+                miningTask.setMiningTask(task);
 
                 // inform player
                 Utils.green(player, "Mining starting... loot every %ss.", settings.getMiningInterval());